scripts/requirements.txt
linkedin-api==2.3.0
PyGithub==2.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pytz==2024.1
